// src/02-functions.ts
import {Friend, Colleague } from './myTypes'
//import objects for testing:
import {friends, colleagues} from './01-basics'

function older(f: Friend) : string {
     f.age += 1
     return `${f.name} is now ${f.age}` 
}

console.log(older(friends[0]))


// Find the colleague with the highest extension number.
function highestExtension(cs: Colleague[]): Colleague {
  const result = cs.sort(
    (c1, c2) => c1.contact.extension - c2.contact.extension
  );
  return result[cs.length - 1];
}
console.log(highestExtension(colleagues.current));


// adds a colleague to an array, and setting their extensins number to the highest extension, plus 1:
function addColleague(cs: Colleague[], name: string, department: string, email: string): String {
  cs[cs.length + 1] = {
  name: name,
  department: department,
  contact: {
    email: email,
    extension: highestExtension(cs).contact.extension + 1,
  },
};
  return 'Colleague added' + name;
}


//to test add colleague:
addColleague(colleagues.current, "Sheild O Connell", "HR", "soc@here.com");
console.log(colleagues.current.filter((c) => c.name === "Sheild O Connell"));

